+++ Flask/app/controllers/ong_routes.py
@@ -1,0 +1,12 @@
+from flask import render_template, Blueprint, redirect, url_for, flash, abort, request, session
+import logging
+from werkzeug.security import generate_password_hash, check_password_hash
+import app.forms.forms as forms 
+from flask_sqlalchemy import SQLAlchemy
+import app.models.models as models
+from app.extensions import db
+from datetime import datetime
+import app.functions as func
+from flask_login import login_user, login_required, logout_user, current_user
+
+ong_routes_bp = Blueprint('ong_routes', __name__)
