+++ Flask/app/templates/landing_page/index.html
@@ -15,6 +15,12 @@
 
 <main class="principal">
     <section class="principal__conteudo">
+        {% for message in get_flashed_messages() %}
+        <div class="alert alert-success alert-dismissible fade show" role="alert">
+            {{ message }}
+            <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+        </div>
+        {% endfor %}
         <div class="principal__noticias">
             <h2 class="principal__noticias-titulo">Notícias da semana</h2>
         </div>
